+++ R/gdalraster.R
@@ -22,11 +22,6 @@
 #' opened dataset, and methods that operate on the dataset as described in 
 #' Details.
 #'
-#' @seealso
-#' Package overview in [`help("gdalraster-package")`][gdalraster-package]
-#'
-#' [create()], [createCopy()], [rasterFromRaster()], [rasterToVRT()], [warp()]
-#'
 #' @section Usage:
 #' \preformatted{
 #' ds <- new(GDALRaster, filename, read_only)
@@ -52,6 +47,7 @@
 #'
 #' ds$getRasterCount()
 #' ds$getBlockSize(band)
+#' ds$getOverviewCount(band)
 #' ds$getDataTypeName(band)
 #' ds$getStatistics(band, approx_ok, force)
 #' ds$getNoDataValue(band)
@@ -97,7 +93,7 @@
 #' Prints various information about the raster dataset to the console (no 
 #' return value, called for that side effect only).
 #' Equivalent to the output of the \command{gdalinfo} command-line utility
-#' (\command{gdalinfo -nomd -norat -noct filename}). Intended here as an 
+#' (\command{gdalinfo -norat -noct filename}). Intended here as an 
 #' informational convenience function.
 #'
 #' \code{$getDriverShortName()}
@@ -173,6 +169,9 @@
 #' be the tile size. Note that the X and Y block sizes don't have to divide 
 #' the image size evenly, meaning that right and bottom edge blocks may be 
 #' incomplete.
+#'
+#' \code{$getOverviewCount(band)}
+#' Returns the number of overview layers available for \code{band}.
 #'
 #' \code{$getDataTypeName(band)}
 #' Returns the name of the pixel data type for \code{band}. The possible data 
@@ -308,8 +307,12 @@
 #' Returns a numeric or complex vector containing the values that were read.
 #' It is organized in left to right, top to bottom pixel order. 
 #' NA will be returned in place of the nodata value if the 
-#' raster dataset has a nodata value defined for this band. No rescaling of the
-#' data is performed (see \code{$getScale()} and \code{$getOffset()} above).
+#' raster dataset has a nodata value defined for this band.
+#' Data are read as R `integer` type when possible for the raster data type
+#' (Byte, Int8, Int16, UInt16, Int32), otherwise as type `double` (UInt32, 
+#' Float32, Float64).
+#' No rescaling of the data is performed (see \code{$getScale()} and 
+#' \code{$getOffset()} above).
 #' An error is raised if the read operation fails.
 #'
 #' \code{$write(band, xoff, yoff, xsize, ysize, rasterData)}
@@ -356,6 +359,23 @@
 #' The dataset can be re-opened on the existing \code{filename} with 
 #' \code{$open(read_only=TRUE)} or \code{$open(read_only=FALSE)}.
 #'
+#' @note
+#' The `$read()` method will perform automatic resampling if the 
+#' specified output size (`out_xsize * out_ysize`) is different than 
+#' the size of the region being read (`xsize * ysize`). In that case, the 
+#' GDAL_RASTERIO_RESAMPLING configuration option could also be defined to 
+#' override the default resampling to one of BILINEAR, CUBIC, CUBICSPLINE, 
+#' LANCZOS, AVERAGE or MODE (see [set_config_option()]).
+#'
+#' @seealso
+#' Package overview in [`help("gdalraster-package")`][gdalraster-package]
+#'
+#' [create()], [createCopy()], [rasterFromRaster()], [rasterToVRT()]
+#'
+#' [fillNodata()], [warp()], [plot_raster()]
+#'
+#' `read_ds()` is a convenience wrapper for `GDALRaster$read()`.
+#'
 #' @examples
 #' lcp_file <- system.file("extdata/storm_lake.lcp", package="gdalraster")
 #' ds <- new(GDALRaster, lcp_file, read_only=TRUE)
@@ -378,6 +398,7 @@
 #' ## retrieve the number of bands and some band-level parameters
 #' ds$getRasterCount()
 #' ds$getBlockSize(band=1)
+#' ds$getOverviewCount(band=1)
 #' ds$getDataTypeName(band=1)
 #' ds$getNoDataValue(band=1)
 #'