+++ src/gdal_exp.cpp
@@ -7,9 +7,11 @@
 #include <cmath>
 
 #include "gdal.h"
-#include "gdal_utils.h"
 #include "cpl_conv.h"
+#include "cpl_port.h"
 #include "cpl_string.h"
+#include "gdal_alg.h"
+#include "gdal_utils.h"
 
 #include <errno.h>
 
@@ -94,6 +96,22 @@ void set_config_option(std::string key, std::string value) {
 	CPLSetConfigOption(key.c_str(), value_);
 }
 
+//' Get the size of memory in use by the GDAL block cache
+//'
+//' `get_cache_used()` returns the amount of memory currently in use for
+//' GDAL block caching. This a wrapper for `GDALGetCacheUsed64()` with return
+//' value as MB.
+//'
+//' @returns Integer. Amount of cache memory in use in MB.
+//'
+//' @examples
+//' get_cache_used()
+// [[Rcpp::export]]
+int get_cache_used() {
+	GIntBig nCacheUsed = GDALGetCacheUsed64();
+	return static_cast<int>(nCacheUsed / (1000 * 1000));
+}
+
 //' Create a new uninitialized raster
 //'
 //' `create()` makes an empty raster in the specified format.
@@ -117,7 +135,8 @@ void set_config_option(std::string key, std::string value) {
 //' [`GDALRaster-class`][GDALRaster], [createCopy()], [rasterFromRaster()]
 //' @examples
 //' new_file <- paste0(tempdir(), "/", "newdata.tif")
-//' create("GTiff", new_file, 143, 107, 1, "Int16")
+//' create(format="GTiff", dst_filename=new_file, xsize=143, ysize=107,
+//'        nbands=1, dataType="Int16")
 //' ds <- new(GDALRaster, new_file, read_only=FALSE)
 //' ## EPSG:26912 - NAD83 / UTM zone 12N
 //' ds$setProjection(epsg_to_wkt(26912))
@@ -196,11 +215,12 @@ bool create(std::string format, std::string dst_filename,
 //' lcp_file <- system.file("extdata/storm_lake.lcp", package="gdalraster")
 //' tif_file <- paste0(tempdir(), "/", "storml_lndscp.tif")
 //' options <- c("COMPRESS=LZW")
-//' createCopy("GTiff", tif_file, lcp_file, options=options)
+//' createCopy(format="GTiff", dst_filename=tif_file, src_filename=lcp_file,
+//'            options=options)
 //' file.size(lcp_file)
 //' file.size(tif_file)
 //' ds <- new(GDALRaster, tif_file, read_only=FALSE)
-//' ds$getMetadata(0, "IMAGE_STRUCTURE")
+//' ds$getMetadata(band=0, domain="IMAGE_STRUCTURE")
 //' for (band in 1:ds$getRasterCount())
 //'     ds$setNoDataValue(band, -9999)
 //' ds$getStatistics(band=1, approx_ok=FALSE, force=TRUE)
@@ -219,6 +239,8 @@ bool createCopy(std::string format, std::string dst_filename,
 		Rcpp::stop("Driver does not support create copy.");
 		
 	GDALDatasetH hSrcDS = GDALOpenShared(src_filename.c_str(), GA_ReadOnly);
+	if(hSrcDS == NULL)
+		Rcpp::stop("Open source raster failed.");
 	
 	std::vector<char *> opt_list = {NULL};
 	if (options.isNotNull()) {
@@ -341,7 +363,7 @@ Rcpp::NumericVector inv_geotransform(const std::vector<double> gt) {
 //' pts <- read.csv(pt_file)
 //' print(pts)
 //' raster_file <- system.file("extdata/storm_lake.lcp", package="gdalraster")
-//' ds <- new(GDALRaster, raster_file, TRUE)
+//' ds <- new(GDALRaster, raster_file, read_only=TRUE)
 //' gt <- ds$getGeoTransform()
 //' get_pixel_line(as.matrix(pts[,-1]), gt)
 //' ds$close()
@@ -366,6 +388,86 @@ Rcpp::IntegerMatrix get_pixel_line(const Rcpp::NumericMatrix xy,
 											inv_gt[5] * geo_y));
 	}
 	return pixel_line;
+}
+
+//' Fill selected pixels by interpolation from surrounding areas
+//'
+//' `fillNodata()` is a wrapper for `GDALFillNodata()` in the GDAL Algorithms
+//' API. This algorithm will interpolate values for all designated nodata 
+//' pixels (pixels having an intrinsic nodata value, or marked by zero-valued
+//' pixels in the optional raster specified in `mask_file`). For each nodata 
+//' pixel, a four direction conic search is done to find values to interpolate
+//' from (using inverse distance weighting).
+//' Once all values are interpolated, zero or more smoothing iterations
+//' (3x3 average filters on interpolated pixels) are applied to smooth out 
+//' artifacts.
+//'
+//' @note
+//' The input raster will be modified in place. It should not be open in a
+//' `GDALRaster` object while processing with `fillNodata()`.
+//'
+//' @param filename Filename of input raster in which to fill nodata pixels.
+//' @param band Integer band number to modify in place.
+//' @param mask_file Optional filename of raster to use as a validity mask
+//' (band 1 is used, zero marks nodata pixels, non-zero marks valid pixels).
+//' @param max_dist Maximum distance (in pixels) that the algorithm 
+//' will search out for values to interpolate (100 pixels by default).
+//' @param smooth_iterations The number of 3x3 average filter smoothing
+//' iterations to run after the interpolation to dampen artifacts
+//' (0 by default).
+//' @returns Logical indicating success (invisible \code{TRUE}).
+//' An error is raised if the operation fails.
+//' @examples
+//' ## fill nodata edge pixels in the elevation raster
+//' elev_file <- system.file("extdata/storml_elev.tif", package="gdalraster")
+//' 
+//' ## get count of nodata
+//' df = combine(elev_file)
+//' head(df)
+//' df[is.na(df$storml_elev),]
+//' 
+//' ## make a copy that will be modified
+//' mod_file <- paste0(tempdir(), "/", "storml_elev_fill.tif")
+//' file.copy(elev_file,  mod_file)
+//' 
+//' fillNodata(mod_file, band=1)
+//' 
+//' df_mod = combine(mod_file)
+//' head(df_mod)
+//' df_mod[is.na(df_mod$storml_elev_fill),]
+// [[Rcpp::export(invisible = true)]]
+bool fillNodata(std::string filename, int band, std::string mask_file = "",
+		double max_dist = 100, int smooth_iterations = 0) {
+
+	GDALDatasetH hDS = NULL;
+	GDALRasterBandH hBand = NULL;
+	GDALDatasetH hMaskDS = NULL;
+	GDALRasterBandH hMaskBand = NULL;
+	CPLErr err;
+	
+	hDS = GDALOpenShared(filename.c_str(), GA_Update);
+	if(hDS == NULL)
+		Rcpp::stop("Open raster failed.");
+	hBand = GDALGetRasterBand(hDS, band);
+	
+	if (mask_file != "") {
+		hMaskDS = GDALOpenShared(mask_file.c_str(), GA_ReadOnly);
+		if(hMaskDS == NULL)
+			Rcpp::stop("Open raster failed.");
+		hMaskBand = GDALGetRasterBand(hMaskDS, 1);
+	}
+	
+	err = GDALFillNodata(hBand, hMaskBand, max_dist, 0, smooth_iterations,
+			NULL, GDALTermProgressR, NULL);
+	
+	if (err != CE_None)
+		Rcpp::stop("fillNodata() failed.");
+		
+	GDALClose(hDS);
+	if (hMaskDS != NULL)
+		GDALClose(hMaskDS);
+		
+	return true;
 }
 
 //' Raster reprojection